Add tests to CI
