+++ pyproject.toml
@@ -19,11 +19,13 @@ dependencies = [
 ]
 
 [project.optional-dependencies]
-dev = [
-    "pytest~=6.0",
+pre-commit = [
     "pre-commit~=2.16.0",
     "mypy==0.931",
     "types-setuptools==57.4.7",
+]
+test = [
+    "pytest~=6.0",
 ]
 
 [project.urls]
